Merged PR 5011: Add GeneratorProcessingOperator
`GeneratorProcessingOperator` is used to take input from operators that returns generators, such as the `IterateCsvOperator` and the `IterateJsonOperator`.

It accepts a function that takes as input a generator and processes it to return some output. This function is then applied to all the generators it receives from the previous operator in the pipeline.

Related work items: #7945